refactor(doc): rename namespace and move DocChannel and DocOptionsNative classes
